requirements.txt
annotated-types==0.7.0
anyio==4.11.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
distro==1.9.0
exceptiongroup==1.3.0
Flask==2.3.3
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
langdetect==1.0.9
MarkupSafe==3.0.3
openai==2.7.1
packaging==25.0
pluggy==1.6.0
pydantic==2.12.4
pydantic_core==2.41.5
pytest==7.4.3
python-dotenv==1.0.0
requests==2.31.0
six==1.17.0
sniffio==1.3.1
tomli==2.3.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
zipp==3.23.0
